stage1/requirements.txt
torch>=2.5.0
torchvision
torchaudio
transformers>=4.40.0
accelerate>=0.30.0
datasets>=2.18.0
pandas>=2.0.0
numpy
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
trl>=0.8.0
peft>=0.10.0
bitsandbytes>=0.43.0
tqdm
jsonlines
sentencepiece
protobuf
pyyaml
